feat: moving generated code from pkg/generated to generated
