+++ templates/dsl/entity.ts.hbs
@@ -1,5 +1,5 @@
 {{#kaplang-render this}}
-//#FILENAME:pkg/generated/entities/{{type name}}.go:write-always
+//#FILENAME:generated/entities/{{type name}}.go:write-always
 //
 // GENERATED SOURCE - DO NOT EDIT
 //
