+++ templates/dsl/rest-api/health.go.hbs
@@ -1,4 +1,4 @@
-//#FILENAME:pkg/generated/rest_api/healthroute.go:write-always
+//#FILENAME:generated/rest_api/healthroute.go:write-always
 //
 // GENERATED SOURCE - DO NOT EDIT
 //
