+++ templates/dsl/rest-api/routes.go.hbs
@@ -1,4 +1,4 @@
-//#FILENAME:pkg/generated/rest_api/{{lowercase (controller-name this)}}_router.go:write-always
+//#FILENAME:generated/rest_api/{{lowercase (controller-name this)}}_router.go:write-always
 //
 // GENERATED SOURCE - DO NOT EDIT
 //
@@ -6,7 +6,7 @@ package rest
 import (
    "github.com/labstack/echo/v4"
    sdkgoconfig "github.com/kapetacom/sdk-go-config/providers"
-   genservices  "{{@root.options.basePackage}}/pkg/generated/services"
+   genservices  "{{@root.options.basePackage}}/generated/services"
    "{{@root.options.basePackage}}/pkg/services"
 )
