+++ templates/dsl/rest-api/services_interface.go.hbs
@@ -1,4 +1,4 @@
-//#FILENAME:pkg/generated/services/{{lowercase (controller-name this)}}_interface.go:write-always
+//#FILENAME:generated/services/{{lowercase (controller-name this)}}_interface.go:write-always
 // GENERATED SOURCE - DO NOT EDIT
 //
 package services
