+++ templates/kapeta/block-type-service/auth-middleware.hbs
@@ -1,5 +1,5 @@
 {{#usesAnyOf (toArray 'kapeta/resource-type-auth-jwt-provider' 'kapeta/resource-type-auth-jwt-consumer') }}
-//#FILENAME:pkg/generated/auth/auth-middleware.go:write-always
+//#FILENAME:generated/auth/auth-middleware.go:write-always
 package auth
 
 import (
