+++ templates/kapeta/block-type-service/config-entity.ts.hbs
@@ -1,5 +1,5 @@
 {{#kaplang-config data.spec.configuration.source}}
-//#FILENAME:pkg/generated/config/{{type name}}Config.go:write-always
+//#FILENAME:generated/config/{{type name}}Config.go:write-always
 package config
 import (
     "fmt"
