+++ templates/kapeta/block-type-service/router.go.hbs
@@ -1,9 +1,9 @@
-//#FILENAME:pkg/generated/router.go:write-always
+//#FILENAME:generated/router.go:write-always
 package generated
 import (
     "github.com/labstack/echo/v4"
 {{#providers-of-type 'kapeta/resource-type-rest-api'}}
-    "{{@root.options.basePackage}}/pkg/generated/rest_api"
+    "{{@root.options.basePackage}}/generated/rest_api"
 {{/providers-of-type}}
     sdkgoconfig "github.com/kapetacom/sdk-go-config/providers"
 )
