+++ templates/kapeta/resource-type-cloud-bucket/client.ts.hbs
@@ -1,4 +1,4 @@
-//#FILENAME:pkg/generated/clients/{{pascalCase data.metadata.name}}Client.go:write-always
+//#FILENAME:generated/clients/{{pascalCase data.metadata.name}}Client.go:write-always
 //
 // GENERATED SOURCE - DO NOT EDIT
 //
