+++ templates/kapeta/resource-type-redis/redis.go.hbs
@@ -1,4 +1,4 @@
-//#FILENAME:pkg/generated/data/{{type data.metadata.name}}DB.go:write-always
+//#FILENAME:generated/data/{{type data.metadata.name}}DB.go:write-always
 //
 // GENERATED SOURCE - DO NOT EDIT
 //
