+++ test/resources/examples/users/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/kapeta/todo/pkg/generated"
+	"github.com/kapeta/todo/generated"
 	sdkgoconfig "github.com/kapetacom/sdk-go-config"
 	"github.com/labstack/echo/v4"
 )
